+++ crates/polars-plan/src/dsl/functions/selectors.rs
@@ -58,7 +58,7 @@ pub fn dtype_cols<DT: AsRef<[DataType]>>(dtype: DT) -> Expr {
 }
 
 /// Select multiple columns by index.
-pub fn index_cols<N: AsRef<[i32]>>(indices: N) -> Expr {
+pub fn index_cols<N: AsRef<[i64]>>(indices: N) -> Expr {
     let indices = indices.as_ref().to_vec();
     Expr::IndexColumn(Arc::from(indices))
 }
